# Optimizer changes for the LIPFilter feature.