There will be three PRs for adding the _LIPFilter_ feature: optimizer changes / execution generator changes / relational operator changes. This is the first PR that does optimizer changes.

LIP (lookahead information passing) filter is an improved semi-join optimization technique that pushes "future selectivities" down a left-deep query plan tree. For linear / bush trees, we can also search for the left-deep pattern in the tree and apply the LIP filter optimization.

With LIP, the overall SSB benchmark (scale factor 100) performance will be improved from ~19 seconds to ~12 seconds on the cloudlab machine. The TPC-H benchmark performance will also be improved greatly (e.g. `Q02 6.7s -> 0.7s, Q17 166s -> 1.7s, Q20 55s -> 1.7s`).

Main changes in this PR:
(1) Rewrote `StarSchemaHashJoinOrderOptimization`. This changed some of the TPC-H query plans. One or two SSB query plans were also slightly changed due to an adjustment in `StarSchemaSimpleCostModel`. Nevertheless we will achieve an overall performance of ~12 seconds for the SSB benchmark when the full feature is there.
(2) Added a `LIPFilterConfiguration` data structure to store the LIP filter information during query optimization.
(3) Changed `ExecutionHeuristics` to an optimizer rule `AttachLIPFilters` to decide and attach LIP filters to the physical query plan.