@pateljm 2016-10-17
@jianqiao Nice! LGTM. If you squash the commits to one and rebase, I can merge.