@jianqiao 2016-10-17
@pateljm Thanks! Will sit together with Harshad tomorrow to go through the code once more. Perhaps it's safer to merge the PR after that.